# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AskQuery": {
        "additionalProperties": false,
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskQuery",
        "type": "object"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "needs_clarification": {
            "default": false,
            "title": "Needs Clarification",
            "type": "boolean"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "scoped_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoped Document Id"
          }
        },
        "required": [
          "question",
          "answer",
          "mode"
        ],
        "title": "AskResponse",
        "type": "object"
      },
...
      "Body_image_search_api_v1_search_image_post": {
        "properties": {
          "image": {
            "contentMediaType": "application/octet-stream",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_image_search_api_v1_search_image_post",
        "type": "object"
      },
...
      "Body_visual_image_search_api_v1_search_visual_image_post": {
        "properties": {
          "image": {
            "contentMediaType": "application/octet-stream",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_visual_image_search_api_v1_search_visual_image_post",
        "type": "object"
      },
      "Candidate": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title"
        ],
        "title": "Candidate",
        "type": "object"
      },
      "Citation": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "index",
          "document_id",
          "title"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkfEntryCreate": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 1048576,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "content"
        ],
        "title": "OkfEntryCreate",
        "type": "object"
      },
...
      "SearchHit": {
        "properties": {
          "doc_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Class"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "match_ranges": {
            "items": {
              "$ref": "#/components/schemas/SearchMatchRange"
            },
            "title": "Match Ranges",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "snippet",
          "score"
        ],
        "title": "SearchHit",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "mode",
          "count",
          "hits"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SemanticQuery": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "q": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Q",
            "type": "string"
          }
        },
        "required": [
          "q"
        ],
        "title": "SemanticQuery",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VisualFeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "enum": [
              "relevant",
              "irrelevant",
              "unauthorized",
              "unsafe",
              "wrong_type"
            ],
            "title": "Category",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 240,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "result_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Result Id",
            "type": "string"
          }
        },
        "required": [
          "result_id",
          "category"
        ],
        "title": "VisualFeedbackRequest",
        "type": "object"
      },
...
      "VisualSearchQuery": {
        "additionalProperties": false,
        "description": "Bounded text-to-visual query contract.",
        "properties": {
          "limit": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "mode": {
            "default": "text_to_image",
            "enum": [
              "text_to_image",
              "text_to_page",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "q": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Q",
            "type": "string"
          }
        },
        "required": [
          "q"
        ],
        "title": "VisualSearchQuery",
        "type": "object"
      },
      "VisualSearchResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/VisualSearchHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "mode": {
            "enum": [
              "text_to_image",
              "text_to_page",
              "image_to_image",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "mode",
          "count",
          "hits",
          "provider"
        ],
        "title": "VisualSearchResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "AI-based Digital Document Management System — local vertical slice.",
    "title": "XENIUS DocVault API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "operationId": "keyword_search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 2000,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Keyword Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/ask": {
      "post": {
        "description": "Ask a natural-language question; get a grounded answer with citations (RAG).\n\nRetrieval is security-trimmed to what the user may VIEW, so the AI answer\nnever draws on documents the user cannot access.",
        "operationId": "ask_api_v1_search_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Ask",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/image": {
      "post": {
        "description": "Process an image query ephemerally; visual indexes remain opt-in.",
        "operationId": "image_search_api_v1_search_image_post",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_image_search_api_v1_search_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Image Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/okf/entries": {
      "get": {
        "description": "List all OKF bundle entries (title, category, tags).",
        "operationId": "list_okf_entries_api_v1_search_okf_entries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Okf Entries",
        "tags": [
          "search",
          "okf"
        ]
      },
      "post": {
        "description": "Add or replace a Markdown entry in the OKF bundle.\n\nThe file must include YAML frontmatter with at minimum a `title` field.\nOn success the bundle is reloaded from disk.",
        "operationId": "create_okf_entry_api_v1_search_okf_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OkfEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Okf Entry",
        "tags": [
          "search",
          "okf"
        ]
      }
    },
    "/api/v1/search/okf/reload": {
      "post": {
        "description": "Force-reload the OKF bundle from disk (useful after manual edits).",
        "operationId": "reload_okf_bundle_api_v1_search_okf_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reload Okf Bundle",
        "tags": [
          "search",
          "okf"
        ]
      }
    },
    "/api/v1/search/okf/status": {
      "get": {
        "description": "Return the status of the Open Knowledge Format bundle.",
        "operationId": "okf_status_api_v1_search_okf_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Okf Status",
        "tags": [
          "search",
          "okf"
        ]
      }
    },
    "/api/v1/search/semantic": {
      "post": {
        "description": "Natural-language search.\n\nWhen transformer embeddings are enabled it would vector-search; in this\nlocal slice it falls back to the FTS keyword index while preserving the\nsame request/response contract (mode reported honestly).",
        "operationId": "semantic_search_api_v1_search_semantic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Semantic Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/visual": {
      "post": {
        "description": "Search authorized image assets or rendered visual document pages by text.",
        "operationId": "visual_search_api_v1_search_visual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualSearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisualSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Visual Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/visual-assets/{asset_id}/preview": {
      "get": {
        "description": "Stream a derivative only after fresh SQL document authorization.",
        "operationId": "visual_asset_preview_api_v1_search_visual_assets__asset_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Visual Asset Preview",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/visual-feedback": {
      "post": {
        "description": "Accept bounded feedback without storing result content or notes.",
        "operationId": "visual_feedback_api_v1_search_visual_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Visual Feedback",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/search/visual/image": {
      "post": {
        "description": "Search semantically similar authorized visual assets/pages.\n\nThe uploaded query image is validated, normalized, embedded in memory, and\ndiscarded.  It is never inserted into a visual asset or vector table.",
        "operationId": "visual_image_search_api_v1_search_visual_image_post",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "image_to_image",
              "enum": [
                "image_to_image",
                "hybrid"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_visual_image_search_api_v1_search_visual_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisualSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Visual Image Search",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}